refactor: linting (#178)
